refactor: コンポーネントファイルのフォルダー構造を機能別に整理しました
