+++ src/components/Recipes/FavoriteRecipes.tsx
@@ -1,10 +1,10 @@
 import { Link } from "@tanstack/react-router";
 import styles from "./FavoriteRecipes.module.css";
-import { useCancelFavRecipes } from "../hooks/useQueryHooks";
-import NoImage from "../assets/NoImg.jpg";
-import { FavRecipes } from "../types";
-import { getBackColorByTime } from "../utils/getBackColorByTime";
-import { getCookingTImeLabel } from "../utils/getCookingTimeLabel";
+import { useCancelFavRecipes } from "@/hooks/useQueryHooks";
+import NoImage from "@/assets/NoImg.jpg";
+import { FavRecipes } from "@/types";
+import { getBackColorByTime } from "@/utils/getBackColorByTime";
+import { getCookingTImeLabel } from "@/utils/getCookingTimeLabel";
 
 const FavoriteRecipes = ({
   favoriteRecipes,
