+++ src/components/Recipes/Recipes.tsx
@@ -1,9 +1,9 @@
 import { Link } from "@tanstack/react-router";
-import { RECIPES } from "../types";
+import { RECIPES } from "@/types";
 import styles from "./Recipes.module.css";
-import NoImage from "../assets/NoImg.jpg";
-import { getBackColorByTime } from "../utils/getBackColorByTime";
-import { getCookingTImeLabel } from "../utils/getCookingTimeLabel";
+import NoImage from "@/assets/NoImg.jpg";
+import { getBackColorByTime } from "@/utils/getBackColorByTime";
+import { getCookingTImeLabel } from "@/utils/getCookingTimeLabel";
 
 const Recipes = ({ recipes }: { recipes: RECIPES["data"] }) => {
   return (
