--- src/components/SkeletonRecipe.module.css
@@ -1,81 +1,0 @@
-.skeleton_wrapper {
-  width: 90%;
-  margin: 1rem auto;
-  padding: 0.5rem;
-  text-align: center;
-  box-shadow: 0 0 4px 0 var(--shadow-color);
-  border-radius: 1rem;
-}
-
-.skeleton_heading {
-  font-size: 1.3rem;
-  width: 70%;
-  margin: 0.5rem auto;
-  background-color: #ededed;
-  border-radius: 2rem;
-  position: relative;
-  overflow: hidden;
-}
-
-.skeleton_heading::after {
-  content: "";
-  position: absolute;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  background-image: linear-gradient(90deg, #ededed, #f7f7f7, #ededed);
-  transform: translateX(-100%);
-  animation: skeletonFlowing 1s infinite;
-}
-
-.skeleton_img {
-  width: 50%;
-  height: 250px;
-  margin: 0 auto 1rem;
-  background-color: #ededed;
-  position: relative;
-  overflow: hidden;
-}
-
-.skeleton_img::after {
-  content: "";
-  position: absolute;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  background-image: linear-gradient(90deg, #ededed, #f7f7f7, #ededed);
-  transform: translateX(-100%);
-  animation: skeletonFlowing 1s infinite;
-}
-
-.skeleton_text {
-  font-size: 1.2rem;
-  text-align: left;
-  width: 90%;
-  margin: auto;
-  display: inline-block;
-  border-radius: 0.5rem;
-  background-color: #ededed;
-  position: relative;
-  overflow: hidden;
-}
-
-.skeleton_text::after {
-  content: "";
-  position: absolute;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  background-image: linear-gradient(90deg, #ededed, #f7f7f7, #ededed);
-  transform: translateX(-100%);
-  animation: skeletonFlowing 1s infinite;
-}
-
-@keyframes skeletonFlowing {
-  100% {
-    transform: translateX(100%);
-  }
-}
