+++ src/routes/$followId/followers.lazy.tsx
@@ -1,11 +1,11 @@
 import { createLazyFileRoute } from "@tanstack/react-router";
-import styles from "../../styles/Follow.module.css";
+import styles from "@/styles/Follow.module.css";
 import {
   useFetchAuthInfo,
   useFetchFollowersByUser,
   useFetchUserInfoByParams,
-} from "../../hooks/useQueryHooks";
-import { FollowersList } from "../../components/UserList/FollowersList";
+} from "@/hooks/useQueryHooks";
+import { FollowersList } from "@/components/Follow/FollowersList";
 
 export const Route = createLazyFileRoute("/$followId/followers")({
   component: Followers,
@@ -14,7 +14,9 @@ export const Route = createLazyFileRoute("/$followId/followers")({
 function Followers() {
   const { useParams } = Route;
   const params = useParams();
-  const { data: paramsUserFollowers } = useFetchFollowersByUser(params.followId);
+  const { data: paramsUserFollowers } = useFetchFollowersByUser(
+    params.followId
+  );
   const { data: myUser } = useFetchAuthInfo();
   const { data: paramsUser } = useFetchUserInfoByParams(params.followId);
   return (
@@ -22,7 +24,10 @@ function Followers() {
       <h2>{paramsUser?.name}</h2>
       <h2>フォロワー</h2>
       {paramsUserFollowers ? (
-        <FollowersList users={paramsUserFollowers} loginUserId={myUser?.user_id} />
+        <FollowersList
+          users={paramsUserFollowers}
+          loginUserId={myUser?.user_id}
+        />
       ) : null}
     </div>
   );
