+++ src/routes/favorites/route.lazy.tsx
@@ -1,9 +1,7 @@
 import { createLazyFileRoute } from "@tanstack/react-router";
-import styles from "./route.module.css"
-import {
-  useFetchFavoritesRecipes,
-} from "../../hooks/useQueryHooks";
-import FavoriteRecipes from "../../components/FavoriteRecipes";
+import styles from "./route.module.css";
+import { useFetchFavoritesRecipes } from "@/hooks/useQueryHooks";
+import FavoriteRecipes from "@/components/Recipes/FavoriteRecipes";
 export const Route = createLazyFileRoute("/favorites")({
   component: Favorites,
 });
